SDCICD-352. Fix machine status comparison.
The machine status comparison was inadvertently checking pointer
addresses instead of values.
